Fully Working DatabaseManager, not yet implemented into Server. Added a debug constant to the Constants file.

+++ Source/App/src/dimappers/android/pub/ChoosePub.java
@@ -81,8 +81,8 @@ public class ChoosePub extends ListActivity implements OnClickListener {
 			 Place place = listItems.get(pos);
 			 event.SetPubLocation(
 					 new PubLocation(
-							 place.geometry.location.lat,
-							 place.geometry.location.lng,
+							 (float)place.geometry.location.lat,
+							 (float)place.geometry.location.lng,
 							 place.name)
 					 );
 			 Place_Detailed pd = null;
